+++ src/main/java/ganymedes01/etfuturum/blocks/itemblocks/ItemBlockDyedBed.java
@@ -16,57 +16,45 @@ public class ItemBlockDyedBed extends ItemBlock {
 		setMaxStackSize(1);
 	}
 
-	public boolean onItemUse(ItemStack p_77648_1_, EntityPlayer p_77648_2_, World p_77648_3_, int p_77648_4_, int p_77648_5_, int p_77648_6_, int p_77648_7_, float p_77648_8_, float p_77648_9_, float p_77648_10_)
-	{
-		if (p_77648_3_.isRemote)
-		{
+	public boolean onItemUse(ItemStack p_77648_1_, EntityPlayer p_77648_2_, World p_77648_3_, int p_77648_4_, int p_77648_5_, int p_77648_6_, int p_77648_7_, float p_77648_8_, float p_77648_9_, float p_77648_10_) {
+		if (p_77648_3_.isRemote) {
 			return true;
-		}
-		else if (p_77648_7_ != 1)
-		{
+		} else if (p_77648_7_ != 1) {
 			return false;
-		}
-		else
-		{
+		} else {
 			++p_77648_5_;
-			BlockBed blockbed = (BlockBed)field_150939_a;
-			int i1 = MathHelper.floor_double((double)(p_77648_2_.rotationYaw * 4.0F / 360.0F) + 0.5D) & 3;
+			BlockBed blockbed = (BlockBed) field_150939_a;
+			int i1 = MathHelper.floor_double((double) (p_77648_2_.rotationYaw * 4.0F / 360.0F) + 0.5D) & 3;
 			byte b0 = 0;
 			byte b1 = 0;
 
-			if (i1 == 0)
-			{
+			if (i1 == 0) {
 				b1 = 1;
 			}
 
-			if (i1 == 1)
-			{
+			if (i1 == 1) {
 				b0 = -1;
 			}
 
-			if (i1 == 2)
-			{
+			if (i1 == 2) {
 				b1 = -1;
 			}
 
-			if (i1 == 3)
-			{
+			if (i1 == 3) {
 				b0 = 1;
 			}
 
-			if (p_77648_2_.canPlayerEdit(p_77648_4_, p_77648_5_, p_77648_6_, p_77648_7_, p_77648_1_) && p_77648_2_.canPlayerEdit(p_77648_4_ + b0, p_77648_5_, p_77648_6_ + b1, p_77648_7_, p_77648_1_))
-			{
-				if (p_77648_3_.isAirBlock(p_77648_4_, p_77648_5_, p_77648_6_) && p_77648_3_.isAirBlock(p_77648_4_ + b0, p_77648_5_, p_77648_6_ + b1) && World.doesBlockHaveSolidTopSurface(p_77648_3_, p_77648_4_, p_77648_5_ - 1, p_77648_6_) && World.doesBlockHaveSolidTopSurface(p_77648_3_, p_77648_4_ + b0, p_77648_5_ - 1, p_77648_6_ + b1))
-				{
+			if (p_77648_2_.canPlayerEdit(p_77648_4_, p_77648_5_, p_77648_6_, p_77648_7_, p_77648_1_) && p_77648_2_.canPlayerEdit(p_77648_4_ + b0, p_77648_5_, p_77648_6_ + b1, p_77648_7_, p_77648_1_)) {
+				if (p_77648_3_.isAirBlock(p_77648_4_, p_77648_5_, p_77648_6_) && p_77648_3_.isAirBlock(p_77648_4_ + b0, p_77648_5_, p_77648_6_ + b1) && World.doesBlockHaveSolidTopSurface(p_77648_3_, p_77648_4_, p_77648_5_ - 1, p_77648_6_) && World.doesBlockHaveSolidTopSurface(p_77648_3_, p_77648_4_ + b0, p_77648_5_ - 1, p_77648_6_ + b1)) {
 					p_77648_3_.setBlock(p_77648_4_, p_77648_5_, p_77648_6_, blockbed, i1, 3);
 
-					if (p_77648_3_.getBlock(p_77648_4_, p_77648_5_, p_77648_6_) == blockbed)
-					{
+					if (p_77648_3_.getBlock(p_77648_4_, p_77648_5_, p_77648_6_) == blockbed) {
 						p_77648_3_.setBlock(p_77648_4_ + b0, p_77648_5_, p_77648_6_ + b1, blockbed, i1 + 8, 3);
 					}
 
 					//Disable the sound for continuity, so it doesn't play when the event-based player would not
-					if(ConfigSounds.fixSilentPlacing)p_77648_3_.playSoundEffect((double)((float)p_77648_4_ + 0.5F), (double)((float)p_77648_5_ + 0.5F), (double)((float)p_77648_6_ + 0.5F), this.field_150939_a.stepSound.func_150496_b(), (this.field_150939_a.stepSound.getVolume() + 1.0F) / 2.0F, this.field_150939_a.stepSound.getPitch() * 0.8F);
+					if (ConfigSounds.fixSilentPlacing)
+						p_77648_3_.playSoundEffect((float) p_77648_4_ + 0.5F, (float) p_77648_5_ + 0.5F, (float) p_77648_6_ + 0.5F, this.field_150939_a.stepSound.func_150496_b(), (this.field_150939_a.stepSound.getVolume() + 1.0F) / 2.0F, this.field_150939_a.stepSound.getPitch() * 0.8F);
 					--p_77648_1_.stackSize;
 					return true;
 				}
